Issue 49: S3-S3 Copier Hive Diff Fix (#51)
In S3 the last modified timestamp function of Hadoop actually returns the current time when the object is a directory (with files it behaves as expected), we use the last modified timestamp to generate a checksum which is part of the Hive Diff comparisons

This meant in S3 the diff was failing when testing folder to folder checksums. We now explicitly set directory last modified times to 0, and use the last modified timestamp function with files only.